setting up Context and display alert
added context logic and structure to the codebase and implemented display alert functionality with context. (appContext is basically a wrapper provider that supply all the store values and actions to the children while reducer takes care of the state based on the type of action. type of action is kept separately in constants folder.